Login/Signup buttons and welcome refactor
