+++ src/css/Comment.css
@@ -7,6 +7,7 @@
 .createCommentDiv{
 margin-top: 20px;
 }
+
 .commentsContainer{
   margin-top: 20px;
   background-color: var(--green);
