+++ src/css/Welcome.css
@@ -13,7 +13,7 @@ h1{
   background-color: var(--green);
   border-radius: 8px;
   border: 1px rgb(13, 13, 13) solid;
-  width: 60%;
+  width: 32%;
   height: 140%;
   margin-top: 8%;
 }
@@ -91,6 +91,7 @@ color : var(--light);
 }
 
 #username, #password{
+  height: 1.2rem;
   background: rgb(32, 32, 32);
   border-radius: 1px;
 }
